Update README.md to include a more detailed description
Eniac is a company that primarily deals with Apple-compatible accessories and is limited to Europe. They want to expand their sales in the Brazilian market. In Brazil, Magist is offering a centralized order management system to connect small and medium-sized stores with the biggest Brazilian marketplaces.
So the data we have explored in this section is a Magist data based on various factors as follows;

Products category either Tech or Non- Tech
Delivery status (if they provide delayed or on-time delivery)
Sellers (income, revenue)